refactor(platform-wallet): move the masternode record model into the library crate
The masternode aggregation (`MasternodeAggregate`, `aggregate_masternodes`,
`MasternodeStatus`, `ListMembership`, the DIP-3 payload decode) lived in
`rs-platform-wallet-ffi`, so the only way to get a wallet's masternodes was
through the C ABI — Android's JNI shim and any other host would have had to
re-implement it. It now lives in `platform_wallet::masternode` as
`MasternodeRecord`, with one library entry point,
`PlatformWalletManager::wallet_masternodes_blocking`, that does what the
FFI list function and the withdrawal path each did on their own:
aggregate the wallet's provider transactions, resolve status against the
DML snapshot, and resolve operator / platform-node key ownership by
derive-and-compare.

`MasternodeEntryFFI` is unchanged on the wire; `masternode_entry_ffi` is
now pure marshalling from the record (`order_index`, `operator_key_index`,
`platform_key_index`, `platform_ownership_checked` ride on the record
instead of being computed at the FFI boundary). `MasternodeRecord` carries
a `source: MasternodeSource` (only `Wallet` today) so records from other
provenances can share the same shape.

The aggregation tests move with the code; the FFI keeps a gating test for
the entry marshalling.
